Use swift native type Float instead of CGFloat

+++ Haneke/Format.swift
@@ -18,7 +18,7 @@ public struct Format {
 
     public let allowUpscaling : Bool
     
-    public let compressionQuality : CGFloat = 1.0
+    public let compressionQuality : Float = 1.0
 
     public let size : CGSize
     
@@ -26,7 +26,7 @@ public struct Format {
     
     public let diskCapacity : UInt64
     
-    public init(_ name : String, diskCapacity : UInt64 = 0, size : CGSize = CGSizeZero, scaleMode : ScaleMode = .None, allowUpscaling: Bool = true, compressionQuality : CGFloat = 1.0) {
+    public init(_ name : String, diskCapacity : UInt64 = 0, size : CGSize = CGSizeZero, scaleMode : ScaleMode = .None, allowUpscaling: Bool = true, compressionQuality : Float = 1.0) {
         self.name = name
         self.diskCapacity = diskCapacity
         self.size = size
